moved combined state related code to its own package
