+++ compiller/src/main/java/com/yheriatovych/reductor/processor/Utils.java
@@ -15,7 +15,7 @@ public class Utils {
         R call(T value);
     }
 
-    interface Func2<T1, T2, R> {
+    public interface Func2<T1, T2, R> {
         R call(T1 value1, T2 value2);
     }
